+++ src/1.Core/RequestResponse/Countries/Commands/Create/CreateCountryCommand.cs
@@ -11,12 +11,12 @@ public sealed class CreateCountryCommand : ICommand<long>, IWebRequest
     public string Code { get; set; } = string.Empty;
     public string CoreId { get; set; } = string.Empty;
 
-    public CreateCountryParameter ToCreateParameter(int priority) => new(Title,
+    public CreateCountryParameter ToCreateParameter(long priority) => new(Title,
                                                                          DisplayTitle,
                                                                          CoreId,
                                                                          Code,
                                                                          priority);
-    public RestoreCountryParameter ToRestoreParameter(int priority) => new(Title,
+    public RestoreCountryParameter ToRestoreParameter(long priority) => new(Title,
                                                                            DisplayTitle,
                                                                            Code,
                                                                            priority);
